+++ src/extension.ts
@@ -1,9 +1,10 @@
 import { commands, ExtensionContext } from 'vscode';
 import { ExtensionManagerPanel } from './panels/ExtensionManagerPanel';
 import { dirname } from 'path';
-import {  IS_DEVELOPMENT_MODE, State } from './utils';
+import { emptyDirPromise, IS_DEVELOPMENT_MODE, State } from './utils';
 
 export function activate(context: ExtensionContext) {
+	
 	let rootPath = "";
 	if(IS_DEVELOPMENT_MODE){
 		rootPath = process.env.USERPROFILE + "\\.vscode\\extensions";