+++ src/test/test.ts
@@ -1,1 +1,1 @@
-console.log("AAAA".toLocaleLowerCase());
+