+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./extensionPackage";
 export * from "./extension";
 export * from "./commonUtil";
 export * from "./panelUtils";
+export * from "./promiseUtils";